Update count in-memory store on successful mount/unmount.
If there was an error in Mount and Unmount operations, the in-memory store was getting updated
before the operations {Mount, Unmount} and before it could save the new state to the disk, it
would error out. The in-memory store and on disk state would then get out of sync.
This PR addresses that.

Signed-off-by: Shishir Mahajan <shishir.mahajan@redhat.com>